refactor: Update metadata in constructMetadata function
The metadata in the constructMetadata function has been updated to improve the title and description of the StrathMall marketplace. The title has been changed to "StrathMall - Your Campus & Community Marketplace" and the description now highlights StrathMall as the ultimate marketplace for premium goods, connecting university sellers with savvy shoppers. This change enhances the branding and positioning of the marketplace.