feat: Generate unique slug for product name in SellerProductForm
This commit modifies the SellerProductForm component to automatically generate a unique slug based on the product name. If the product name is provided, the slug is generated using the slugify library. The code then checks if the generated slug already exists in the database using the checkSlugExists function. If the slug is not unique, a random number is appended to the slug and checked again. Finally, the generated slug is set as the value for the "slug" form field.