feat: Add resetToken and resetTokenExpires columns to the "user" table to be used byusers to reset their passwords
This commit adds two new columns, "resetToken" and "resetTokenExpires", to the "user" table in the database. These columns will be used for password reset functionality. The "resetToken" column stores the reset token generated for a user, while the "resetTokenExpires" column stores the expiration date and time for the reset token.