feat: Add discounts array and discount column to product schema
This commit adds a new array of discounts to the product schema in the `discount.ts` file. It also adds a new column named `discount` to the `product` table in the database schema. The `discount` column is of type `numeric(5, 2)` and has a default value of '0'. These changes enable the application to handle discounts for products.